requirements.txt
python-telegram-bot[all]==21.5
google-generativeai==0.8.2
gtts==2.5.1
apscheduler==3.10.4
pytz==2024.1
cachetools==5.3.3
redis==5.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
psutil==5.9.6
asyncpg==0.29.0
aiogram==3.12.0
matplotlib==3.8.2
pillow==10.1.0
qrcode==7.4.2
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
soundfile==0.12.1
pydub==0.25.1
speech-recognition==3.10.0
openai-whisper==20231117 